feat(web): #17 keep working on homepage and views
