+++ web/next.config.js_Zone.Identifier
@@ -1,0 +1,4 @@
+[ZoneTransfer]
+ZoneId=3
+ReferrerUrl=https://drive.google.com/
+HostUrl=https://doc-04-4s-docs.googleusercontent.com/docs/securesc/l9jtkin5340179vc945vldtacantv3ff/qm6b71aadthrlu3b5rtdoi99vr2it4ci/1641567000000/05998778505511157441/09870914982048683453/1X7l0oug4cTCm9lqqVzfe0ckKZ8v7-qBK?e=download&authuser=0&nonce=jpe8mooblov0m&user=09870914982048683453&hash=fjflfvpq5v1n9o08sacr46umdpldf3ug
