+++ web/public/index.html
@@ -6,8 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no" />
     <meta name="theme-color" content="#000000" />
     <meta name="author" content="Devias IO" />
-    <meta name="description" content="A professional kit that comes with ready-to-use Material-UI© components developed with one common goal in mind, help you build faster &amp; beautiful applications. Each component is fully customizable, responsive and easy to integrate." />
-    <meta name="keywords" content="react,material,kit,application,dashboard,admin,template" />
+    <meta name="description" content="A professional site for developpers who wants control over their online services" />
     <link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico" />
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
