+++ web/src/components/dashboard/scaleway/customer-basic-details.tsx
