+++ web/src/components/dashboard/scaleway/customer-emails-summary.tsx
