+++ web/src/components/severity-pill.tsx
@@ -27,32 +27,34 @@ interface SeverityPillRootProps {
 
 const SeverityPillRoot = styled('span')<SeverityPillRootProps>(
   ({ theme, ownerState }) => {
-    const backgroundColor = theme.palette[ownerState.color].main;
-    const color = theme.palette[ownerState.color].contrastText;
+
+    // const backgroundColor = theme.palette[ownerState.color].main;
+    // const color = theme.palette[ownerState.color].contrastText;
+    const backgroundColor = '#14B8A6';
+    const color = 'white';
 
     return {
-      alignItems: 'center',
-      backgroundColor,
-      borderRadius: 12,
-      color,
-      cursor: 'default',
-      display: 'inline-flex',
-      flexGrow: 0,
-      flexShrink: 0,
-      fontFamily: theme.typography.fontFamily,
-      fontSize: theme.typography.pxToRem(12),
-      lineHeight: 2,
-      fontWeight: 600,
-      justifyContent: 'center',
-      letterSpacing: 0.5,
-      minWidth: 20,
-      paddingLeft: theme.spacing(1),
-      paddingRight: theme.spacing(1),
-      textTransform: 'uppercase',
-      whiteSpace: 'nowrap'
-    };
+    alignItems: 'center',
+    backgroundColor,
+    borderRadius: 12,
+    color,
+    cursor: 'default',
+    display: 'inline-flex',
+    flexGrow: 0,
+    flexShrink: 0,
+    fontFamily: theme.typography.fontFamily,
+    fontSize: theme.typography.pxToRem(12),
+    lineHeight: 2,
+    fontWeight: 600,
+    justifyContent: 'center',
+    letterSpacing: 0.5,
+    minWidth: 20,
+    paddingLeft: theme.spacing(1),
+    paddingRight: theme.spacing(1),
+    textTransform: 'uppercase',
+    whiteSpace: 'nowrap'
   }
-);
+});
 
 export const SeverityPill: FC<SeverityPillProps> = (props) => {
   const { color = 'primary', children, ...other } = props;
