+++ web/src/pages/dashboard/AWS/index.tsx
@@ -6,9 +6,9 @@ import { Box, Button, Card, Container, Grid, Typography } from '@mui/material';
 import { productApi } from '../../../__fake-api__/product-api';
 import { AuthGuard } from '../../../components/authentication/auth-guard';
 import { DashboardLayout } from '../../../components/dashboard/dashboard-layout';
-import { ProjectListFilters } from '../../../components/dashboard/product/product-list-filters';
-import type { Filters } from '../../../components/dashboard/product/product-list-filters';
-import { ProductListTable } from '../../../components/dashboard/product/product-list-table';
+import { ProjectListFilters } from '../../../components/dashboard/AWS/product-list-filters';
+import type { Filters } from '../../../components/dashboard/AWS/product-list-filters';
+import { ProductListTable } from '../../../components/dashboard/AWS/product-list-table';
 import { useMounted } from '../../../hooks/use-mounted';
 import { Download as DownloadIcon } from '../../../icons/download';
 import { Upload as UploadIcon } from '../../../icons/upload';
