+++ web/src/pages/dashboard/azure/[orderId].tsx
@@ -8,9 +8,9 @@ import ArrowBackIcon from '@mui/icons-material/ArrowBack';
 import { orderApi } from '../../../__fake-api__/order-api';
 import { AuthGuard } from '../../../components/authentication/auth-guard';
 import { DashboardLayout } from '../../../components/dashboard/dashboard-layout';
-import { OrderItems } from '../../../components/dashboard/order/order-items';
-import { OrderLogs } from '../../../components/dashboard/order/order-logs';
-import { OrderSummary } from '../../../components/dashboard/order/order-summary';
+import { OrderItems } from '../../../components/dashboard/azure/order-items';
+import { OrderLogs } from '../../../components/dashboard/azure/order-logs';
+import { OrderSummary } from '../../../components/dashboard/azure/order-summary';
 import { useMounted } from '../../../hooks/use-mounted';
 import { Calendar as CalendarIcon } from '../../../icons/calendar';
 import { ChevronDown as ChevronDownIcon } from '../../../icons/chevron-down';
