--- web/src/pages/dashboard/customers/index.tsx
@@ -1,396 +1,0 @@
-import { useState, useEffect, useCallback, FormEvent, useRef } from 'react';
-import type { ChangeEvent, MouseEvent } from 'react';
-import type { NextPage } from 'next';
-import Head from 'next/head';
-import {
-  Box,
-  Button,
-  Card,
-  Container,
-  Divider,
-  Grid,
-  InputAdornment,
-  Tab,
-  Tabs,
-  TextField,
-  Typography
-} from '@mui/material';
-import { customerApi } from '../../../__fake-api__/customer-api';
-import { AuthGuard } from '../../../components/authentication/auth-guard';
-import { DashboardLayout } from '../../../components/dashboard/dashboard-layout';
-import { CustomerListTable } from '../../../components/dashboard/customer/customer-list-table';
-import { useMounted } from '../../../hooks/use-mounted';
-import { Download as DownloadIcon } from '../../../icons/download';
-import { Plus as PlusIcon } from '../../../icons/plus';
-import { Search as SearchIcon } from '../../../icons/search';
-import { Upload as UploadIcon } from '../../../icons/upload';
-import { gtm } from '../../../lib/gtm';
-import type { Customer } from '../../../types/customer';
-
-interface Filters {
-  query?: string;
-  hasAcceptedMarketing?: boolean;
-  isProspect?: boolean;
-  isReturning?: boolean;
-}
-
-type Sort =
-  | 'updatedAt|desc'
-  | 'updatedAt|asc'
-  | 'orders|desc'
-  | 'orders|asc';
-
-interface SortOption {
-  value: Sort;
-  label: string;
-}
-
-const tabs = [
-  {
-    label: 'All',
-    value: 'all'
-  },
-  {
-    label: 'Accepts Marketing',
-    value: 'hasAcceptedMarketing'
-  },
-  {
-    label: 'Prospect',
-    value: 'isProspect'
-  },
-  {
-    label: 'Returning',
-    value: 'isReturning'
-  }
-];
-
-const sortOptions: SortOption[] = [
-  {
-    label: 'Last update (newest)',
-    value: 'updatedAt|desc'
-  },
-  {
-    label: 'Last update (oldest)',
-    value: 'updatedAt|asc'
-  },
-  {
-    label: 'Total orders (highest)',
-    value: 'orders|desc'
-  },
-  {
-    label: 'Total orders (lowest)',
-    value: 'orders|asc'
-  }
-];
-
-const applyFilters = (
-  customers: Customer[],
-  filters: Filters
-): Customer[] => customers.filter((customer) => {
-  if (filters.query) {
-    let queryMatched = false;
-    const properties = ['email', 'name'];
-
-    properties.forEach((property) => {
-      if (customer[property].toLowerCase().includes(filters.query.toLowerCase())) {
-        queryMatched = true;
-      }
-    });
-
-    if (!queryMatched) {
-      return false;
-    }
-  }
-
-  if (filters.hasAcceptedMarketing && !customer.hasAcceptedMarketing) {
-    return false;
-  }
-
-  if (filters.isProspect && !customer.isProspect) {
-    return false;
-  }
-
-  if (filters.isReturning && !customer.isReturning) {
-    return false;
-  }
-
-  return true;
-});
-
-const descendingComparator = (
-  a: Customer,
-  b: Customer,
-  orderBy: string
-): number => {
-  if (b[orderBy] < a[orderBy]) {
-    return -1;
-  }
-
-  if (b[orderBy] > a[orderBy]) {
-    return 1;
-  }
-
-  return 0;
-};
-
-const getComparator = (order: 'asc' | 'desc', orderBy: string) => (
-  order === 'desc'
-    ? (a: Customer, b: Customer) => descendingComparator(a, b, orderBy)
-    : (a: Customer, b: Customer) => -descendingComparator(a, b, orderBy)
-);
-
-const applySort = (customers: Customer[], sort: Sort): Customer[] => {
-  const [orderBy, order] = sort.split('|') as [string, 'asc' | 'desc'];
-  const comparator = getComparator(order, orderBy);
-  const stabilizedThis = customers.map((el, index) => [el, index]);
-
-  stabilizedThis.sort((a, b) => {
-    // @ts-ignore
-    const newOrder = comparator(a[0], b[0]);
-
-    if (newOrder !== 0) {
-      return newOrder;
-    }
-
-    // @ts-ignore
-    return a[1] - b[1];
-  });
-
-  // @ts-ignore
-  return stabilizedThis.map((el) => el[0]);
-};
-
-const applyPagination = (
-  customers: Customer[],
-  page: number,
-  rowsPerPage: number
-): Customer[] => customers.slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage);
-
-const CustomerList: NextPage = () => {
-  const isMounted = useMounted();
-  const queryRef = useRef<HTMLInputElement | null>(null);
-  const [customers, setCustomers] = useState<Customer[]>([]);
-  const [currentTab, setCurrentTab] = useState<string>('all');
-  const [page, setPage] = useState<number>(0);
-  const [rowsPerPage, setRowsPerPage] = useState<number>(10);
-  const [sort, setSort] = useState<Sort>(sortOptions[0].value);
-  const [filters, setFilters] = useState<Filters>({
-    query: '',
-    hasAcceptedMarketing: null,
-    isProspect: null,
-    isReturning: null
-  });
-
-  useEffect(() => {
-    gtm.push({ event: 'page_view' });
-  }, []);
-
-  const getCustomers = useCallback(async () => {
-    try {
-      const data = await customerApi.getCustomers();
-
-      if (isMounted()) {
-        setCustomers(data);
-      }
-    } catch (err) {
-      console.error(err);
-    }
-  }, [isMounted]);
-
-  useEffect(
-    () => {
-      getCustomers();
-    },
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    []
-  );
-
-  const handleTabsChange = (event: ChangeEvent<{}>, value: string): void => {
-    const updatedFilters = {
-      ...filters,
-      hasAcceptedMarketing: null,
-      isProspect: null,
-      isReturning: null
-    };
-
-    if (value !== 'all') {
-      updatedFilters[value] = true;
-    }
-
-    setFilters(updatedFilters);
-    setCurrentTab(value);
-  };
-
-  const handleQueryChange = (event: FormEvent<HTMLFormElement>): void => {
-    event.preventDefault();
-    setFilters((prevState) => ({
-      ...prevState,
-      query: queryRef.current?.value
-    }));
-  };
-
-  const handleSortChange = (event: ChangeEvent<HTMLInputElement>): void => {
-    setSort(event.target.value as Sort);
-  };
-
-  const handlePageChange = (event: MouseEvent<HTMLButtonElement> | null, newPage: number): void => {
-    setPage(newPage);
-  };
-
-  const handleRowsPerPageChange = (event: ChangeEvent<HTMLInputElement>): void => {
-    setRowsPerPage(parseInt(event.target.value, 10));
-  };
-
-  // Usually query is done on backend with indexing solutions
-  const filteredCustomers = applyFilters(customers, filters);
-  const sortedCustomers = applySort(filteredCustomers, sort);
-  const paginatedCustomers = applyPagination(sortedCustomers, page, rowsPerPage);
-
-  return (
-    <>
-      <Head>
-        <title>
-          Dashboard: Customer List | Material Kit Pro
-        </title>
-      </Head>
-      <Box
-        component="main"
-        sx={{
-          flexGrow: 1,
-          py: 8
-        }}
-      >
-        <Container maxWidth="xl">
-          <Box sx={{ mb: 4 }}>
-            <Grid
-              container
-              justifyContent="space-between"
-              spacing={3}
-            >
-              <Grid item>
-                <Typography variant="h4">
-                  Customers
-                </Typography>
-              </Grid>
-              <Grid item>
-                <Button
-                  startIcon={<PlusIcon fontSize="small" />}
-                  variant="contained"
-                >
-                  Add
-                </Button>
-              </Grid>
-            </Grid>
-            <Box
-              sx={{
-                m: -1,
-                mt: 3
-              }}
-            >
-              <Button
-                startIcon={<UploadIcon fontSize="small" />}
-                sx={{ m: 1 }}
-              >
-                Import
-              </Button>
-              <Button
-                startIcon={<DownloadIcon fontSize="small" />}
-                sx={{ m: 1 }}
-              >
-                Export
-              </Button>
-            </Box>
-          </Box>
-          <Card>
-            <Tabs
-              indicatorColor="primary"
-              onChange={handleTabsChange}
-              scrollButtons="auto"
-              sx={{ px: 3 }}
-              textColor="primary"
-              value={currentTab}
-              variant="scrollable"
-            >
-              {tabs.map((tab) => (
-                <Tab
-                  key={tab.value}
-                  label={tab.label}
-                  value={tab.value}
-                />
-              ))}
-            </Tabs>
-            <Divider />
-            <Box
-              sx={{
-                alignItems: 'center',
-                display: 'flex',
-                flexWrap: 'wrap',
-                m: -1.5,
-                p: 3
-              }}
-            >
-              <Box
-                component="form"
-                onSubmit={handleQueryChange}
-                sx={{
-                  flexGrow: 1,
-                  m: 1.5
-                }}
-              >
-                <TextField
-                  defaultValue=""
-                  fullWidth
-                  inputProps={{ ref: queryRef }}
-                  InputProps={{
-                    startAdornment: (
-                      <InputAdornment position="start">
-                        <SearchIcon fontSize="small" />
-                      </InputAdornment>
-                    )
-                  }}
-                  placeholder="Search customers"
-                />
-              </Box>
-              <TextField
-                label="Sort By"
-                name="sort"
-                onChange={handleSortChange}
-                select
-                SelectProps={{ native: true }}
-                sx={{ m: 1.5 }}
-                value={sort}
-              >
-                {sortOptions.map((option) => (
-                  <option
-                    key={option.value}
-                    value={option.value}
-                  >
-                    {option.label}
-                  </option>
-                ))}
-              </TextField>
-            </Box>
-            <CustomerListTable
-              customers={paginatedCustomers}
-              customersCount={filteredCustomers.length}
-              onPageChange={handlePageChange}
-              onRowsPerPageChange={handleRowsPerPageChange}
-              rowsPerPage={rowsPerPage}
-              page={page}
-            />
-          </Card>
-        </Container>
-      </Box>
-    </>
-  );
-};
-
-CustomerList.getLayout = (page) => (
-  <AuthGuard>
-    <DashboardLayout>
-      {page}
-    </DashboardLayout>
-  </AuthGuard>
-);
-
-export default CustomerList;
