+++ web/src/pages/index.tsx
@@ -1,13 +1,9 @@
-import { useEffect, useState } from 'react';
 import type { NextPage } from 'next';
 import Head from 'next/head';
 import {
   Box,
-  Button,
   Container,
   Grid,
-  MenuItem,
-  TextField,
   Typography
 } from '@mui/material';
 import { AuthGuard } from '../components/authentication/auth-guard';
@@ -17,8 +13,6 @@ import { OverviewLatestTransactions } from '../components/dashboard/overview/ove
 import { OverviewPrivateWallet } from '../components/dashboard/overview/overview-private-wallet';
 import { OverviewTotalBalance } from '../components/dashboard/overview/overview-total-balance';
 import { OverviewTotalTransactions } from '../components/dashboard/overview/overview-total-transactions';
-
-import { Reports as ReportsIcon } from '../icons/reports';
 
 const Overview: NextPage = () => {
   return (
@@ -46,39 +40,6 @@ const Overview: NextPage = () => {
                 <Typography variant="h4">
                   Good Morning
                 </Typography>
-              </Grid>
-              <Grid
-                item
-                sx={{
-                  display: 'flex',
-                  alignItems: 'center',
-                  m: -1
-                }}
-              >
-                <Button
-                  startIcon={<ReportsIcon fontSize="small" />}
-                  sx={{ m: 1 }}
-                  variant="outlined"
-                >
-                  Reports
-                </Button>
-                <TextField
-                  defaultValue="week"
-                  label="Period"
-                  select
-                  size="small"
-                  sx={{ m: 1 }}
-                >
-                  <MenuItem value="week">
-                    Last week
-                  </MenuItem>
-                  <MenuItem value="month">
-                    Last month
-                  </MenuItem>
-                  <MenuItem value="year">
-                    Last year
-                  </MenuItem>
-                </TextField>
               </Grid>
             </Grid>
           </Box>
